Large refactor move file system intellisense and test into own projects
