+++ Directory.Packages.Props
@@ -43,7 +43,6 @@
     <PackageVersion Include="System.Text.Encoding.CodePages" Version="9.0.2" />
     <PackageVersion Include="T-Digest.NET" Version="1.0.3" />
     <PackageVersion Include="TestableIO.System.IO.Abstractions" Version="22.0.11" />
-    <PackageVersion Include="TestableIO.System.IO.Abstractions.TestingHelpers" Version="22.0.11" />
     <PackageVersion Include="TestableIO.System.IO.Abstractions.Wrappers" Version="22.0.11" />
     <PackageVersion Include="VegaGenerator" Version="1.0.8" />
     <PackageVersion Include="WPF-UI" Version="4.0.2" />
