+++ applications/lokqlDx/MainWindow.xaml.cs
@@ -8,7 +8,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shell;
-using KustoLoco.Core.Intellisense;
+using Intellisense.FileSystem;
 using Lokql.Engine;
 using Microsoft.Win32;
 using NotNullStrings;
