+++ libraries/KustoLoco.Core/KustoLoco.Core.csproj
@@ -39,8 +39,6 @@
     <PackageReference Include="System.Linq.Async" />
     <PackageReference Include="System.Text.Json" />
     <PackageReference Include="T-Digest.NET" />
-    <PackageReference Include="TestableIO.System.IO.Abstractions" />
-    <PackageReference Include="TestableIO.System.IO.Abstractions.Wrappers" />
   </ItemGroup>
 
   <ItemGroup>
